moves nucleus function to assets location
